news-personalization/backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database (PostgreSQL)
sqlalchemy==2.0.30
asyncpg==0.29.0              # async PostgreSQL driver
alembic==1.13.1              # DB migrations

# Redis client
redis==5.0.4

# Celery (background tasks)
celery==5.3.6
celery[redis]==5.3.6

# Password hashing
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

# JWT tokens (for auth)
python-jose[cryptography]==3.3.0

# Environment variables
python-dotenv==1.0.1

# RSS feed parser (for fetching ET news)
feedparser==6.0.11

# HTTP client (for API calls)
httpx==0.27.0

# Google Gemini (FREE LLM)
google-generativeai==0.5.4

# Data validation
pydantic==2.7.1
pydantic-settings==2.2.1

# Date utilities
python-dateutil==2.9.0

#groq
groq